handling merge diffierences
